Using media live's scheduler capability !

+++ src/model/model.go
@@ -10,7 +10,7 @@ type Opinion struct {
 	EndTime			 time.Time   `json:"end_time"`
 	Values           interface{} `json:"values"`
 	ServiceUrl       string      `json:"service_url"`
-	SegmentLength    uint64       `json:"segment_length,omitempty"`
+	SegmentLength    int       `json:"segment_length,omitempty"`
 	Partner			 string      `json:"partner,omitempty"`
 	Deployment       string      `json:"deployment,omitempty"`
 	EventId			 string      `json:"event_id,omitempty"`
